+++ spec/helpers/application_helper_spec.rb
@@ -22,14 +22,36 @@ describe ApplicationHelper do
       page_title("Some Title", show_error: false)
       title = content_for(:page_title)
 
-      expect(title).to eq("Some Title - #{I18n.t("student_loans.journey_name")} - GOV.UK")
+      expect(title).to eq("Some Title - #{I18n.t("student_loans.policy_name")} - GOV.UK")
     end
 
     it "Returns an error prefix" do
       page_title("Some Title", show_error: true)
       title = content_for(:page_title)
 
-      expect(title).to eq("Error - Some Title - #{I18n.t("student_loans.journey_name")} - GOV.UK")
+      expect(title).to eq("Error - Some Title - #{I18n.t("student_loans.policy_name")} - GOV.UK")
+    end
+  end
+
+  describe "#support_email_address" do
+    it "defaults to the generic support address" do
+      expect(support_email_address).to eq t("support_email_address")
+    end
+
+    it "returns a policy-specific email address" do
+      expect(support_email_address("student-loans")).to eq t("student_loans.support_email_address")
+      expect(support_email_address("maths-and-physics")).to eq t("maths_and_physics.support_email_address")
+    end
+  end
+
+  describe "#policy_service_name" do
+    it "defaults to the generic service name" do
+      expect(policy_service_name).to eq t("service_name")
+    end
+
+    it "returns a policy-specific service name" do
+      expect(policy_service_name("student-loans")).to eq t("student_loans.policy_name")
+      expect(policy_service_name("maths-and-physics")).to eq t("maths_and_physics.policy_name")
     end
   end
 end